+++ src/ms3d-acc/StructMS3D.h
@@ -249,7 +249,7 @@ public:
 	float * pVertexArrayRaw; // 最原始坐标，未经坐标变换展开
 	float * pVertexArrayStatic; // 静态
 	float * pVertexArrayDynamic; // 动态
-	int * pIndexJoint;
+	float * pIndexJoint;
 	float * pWeightJoint;
 
 	float * pTexcoord; // 纹理坐标
